Add forceNoserverino flag for operators (#107)
We have seen a large deployment in which upgrading from xenial to jammy
stemcells caused all apps using SMB mounts to fail with "Stale file handle"
errors. This turned out to be because the SMB server was suggesting inode
numbers, instead of allowing the client to generate temporary inode numbers.

The fix was to re-bind the SMB service with the mount parameter
"noserverino". This new flag is intended to allow the platform operator
to apply that fix across the whole foundation, rather than relying on
application authors to re-bind their SMB services.

[#184474526]