Allow nodfs mountoption (#234)
* allow nodfs option

[#186402045]

dealing with a support case it was found that on newer kernels the mounting
of some oracle zfs based shares via samba will fail without the nodfs option.

this can be addressed by whitelisting the nodfs options to be used in the
underlying service bindings.

* add tests for nodfs
[#186402045]
we want to ensure it's set without a value because it doesn't expect a value