switch to ginkgo v2 and adjust for new version of docker integration tests
[#186806700]

the upstream repo for docker driver integration tests removed the vendor directory
and updated some dependencies. this creates two issues for running these tests
against smbdriver.

- the upstream PR is broken
it seems to miss a commit because the included go.sum file is apparently missing contents
these are fixed in a currently unmerged PR
additionally the upstream PR removed the vendor directory while our script tries to build
with the -mod flags set to `vendor`. this needs to be removed to adjust for the removal
of the vendor dir

- we're installing ginkgo v1 in this script and the upstream PR also changed the dependencies
for docker driver integration tests to pull ginkfo v2. remove the go install and replace it
with a go run instead to keep up to date.